requirements.txt
asgiref==3.7.2
Django==3.2.22
django-filter==23.5
djangorestframework==3.14.0
importlib-metadata==7.0.0
Markdown==3.5.1
psycopg2-binary==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
zipp==3.17.0
